requirements.txt
openai>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
loguru>=0.7.0
PyYAML>=6.0.1
requests>=2.31.0
aiohttp>=3.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
alembic>=1.12.0
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0
replicate>=0.22.0
certifi>=2024.2.0 